Resolvendo lint (quase la)
